backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
anthropic==0.40.0
google-genai>=1.68.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
